changes at user add del mod stracture (simplified)
